fix: move `nvim` inside terminal/ and remove redundant API wrappers
`nvim` may be too common of a module name to have it global available, this cause problems with
users that wrote their own `nvim` module or with any other plugin using the same name.

`vim` now bundles with most of the `nvim` APIs including an autocmd API although some wrappers
may still be needed for `nvim-terminal.lua` to work in older versions other than the latest stable
